requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langchain-core>=0.1.0
pandas>=1.5.0
numpy>=1.21.0
